+++ src/main/java/com/project/petcarepedia/controller/MypageController.java
@@ -218,11 +218,4 @@ public class MypageController {
         return viewName;
     }
 
-    @GetMapping("pass_check/{mid}/{pass}")
-    public String pass_check(@PathVariable String mid, @PathVariable String pass) {
-        MemberDto memberDto = new MemberDto();
-        memberDto.setMid(mid);
-        memberDto.setPass(pass);
-        return String.valueOf(memberService.checkPass(memberDto));
-    }
 }
